pyproject.toml
[tool.poetry]
name = "dd"
version = "0.1.0"
description = ""
authors = ["user <ivan.isaev7@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.4"
psycopg2-binary = "^2.9.10"
django-widget-tweaks = "^1.5.0"
utils = "^1.0.2"
requests = "^2.32.4"
djangorestframework = "^3.16.1"
python-telegram-bot = "^22.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
